physics_swarm/requirements.txt
# Core CrewAI Framework
crewai>=0.28.0
crewai-tools>=0.1.0

# LLM APIs
openai>=1.13.3
anthropic>=0.18.0

# Web Research & Data Collection
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.3
selenium>=4.18.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# Database & Storage
sqlalchemy>=2.0.25
alembic>=1.13.0

# Data Processing & Analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
rich>=13.7.0

# Optional: BrightData (if needed)
# brightdata-api==1.0.0  # Uncomment if using BrightData

# Development Tools
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
